+++ tests/integration/aws/route53client_test.go
@@ -23,18 +23,18 @@ var _ = Describe("Route53 Resolver client", func() {
 	})
 
 	When("creating hosted zones", func() {
-		var hostedZoneId string
 		tags := map[string]string{
 			"Name":      "jose",
 			"something": "else",
 		}
 
-		AfterEach(func() {
-			_, err = rawRoute53Client.DeleteHostedZoneWithContext(ctx, &route53.DeleteHostedZoneInput{Id: awssdk.String(hostedZoneId)})
-			Expect(err).NotTo(HaveOccurred())
-		})
-
 		Context("we want a public hosted zone", func() {
+			var hostedZoneId string
+
+			AfterEach(func() {
+				_, _ = rawRoute53Client.DeleteHostedZoneWithContext(ctx, &route53.DeleteHostedZoneInput{Id: awssdk.String(hostedZoneId)})
+			})
+
 			It("creates a public hosted zone successfully", func() {
 				hostedZoneId, err = route53Client.CreateHostedZone(ctx, logger, resolver.BuildPublicHostedZone("apublic.test.example.com", tags))
 				Expect(err).NotTo(HaveOccurred())
@@ -89,16 +89,15 @@ var _ = Describe("Route53 Resolver client", func() {
 		})
 
 		Context("we want a private hosted zone", func() {
-			var cluster resolver.Cluster
+			var hostedZoneId string
 			var dnsZone resolver.DnsZone
 
 			BeforeEach(func() {
-				cluster = resolver.Cluster{
-					Name:   "aprivate",
-					Region: Region,
-					VPCId:  VPCId,
-				}
-				dnsZone = resolver.BuildPrivateHostedZone("aprivate.test.example.com", cluster, tags, []string{MCVPCId})
+				dnsZone = resolver.BuildPrivateHostedZone("aprivate.test.example.com", tags, VPCId, Region, []string{MCVPCId})
+			})
+
+			AfterEach(func() {
+				_, _ = rawRoute53Client.DeleteHostedZoneWithContext(ctx, &route53.DeleteHostedZoneInput{Id: awssdk.String(hostedZoneId)})
 			})
 
 			It("creates a private hosted zone successfully", func() {
@@ -106,13 +105,13 @@ var _ = Describe("Route53 Resolver client", func() {
 				Expect(err).NotTo(HaveOccurred())
 
 				var privateHostedZoneResponse *route53.ListHostedZonesByNameOutput
-				Eventually(func() (int, error) {
+				Eventually(func() (string, error) {
 					privateHostedZoneResponse, err = rawRoute53Client.ListHostedZonesByNameWithContext(ctx, &route53.ListHostedZonesByNameInput{
-						DNSName:  awssdk.String("aprivate.test.example.com"),
+						DNSName:  awssdk.String("aprivate.test.example.com."),
 						MaxItems: awssdk.String("1"),
 					})
-					return len(privateHostedZoneResponse.HostedZones), err
-				}, "2s", "100ms").Should(Equal(1))
+					return *privateHostedZoneResponse.HostedZones[0].Name, err
+				}, "3s", "100ms").Should(Equal("aprivate.test.example.com."))
 				Expect(*privateHostedZoneResponse.HostedZones[0].Name).To(Equal("aprivate.test.example.com."))
 
 				actualTags, err := rawRoute53Client.ListTagsForResourceWithContext(ctx, &route53.ListTagsForResourceInput{
@@ -167,7 +166,7 @@ var _ = Describe("Route53 Resolver client", func() {
 						MaxItems: awssdk.String("1"),
 					})
 					Expect(err).NotTo(HaveOccurred())
-					Expect(len(hostedZoneResponse.HostedZones)).To(Equal(1))
+					Expect(*hostedZoneResponse.HostedZones[0].Name).To(Equal("already.private.exists.test.example.com."))
 				})
 			})
 		})
@@ -320,6 +319,7 @@ var _ = Describe("Route53 Resolver client", func() {
 			})
 			Expect(err).NotTo(HaveOccurred())
 
+			// We add the delegation so we can delete it later.
 			err = route53Client.AddDelegationToParentZone(ctx, logger, *parentHostedZoneToFind.HostedZone.Id, *hostedZoneToFind.HostedZone.Id)
 			Expect(err).NotTo(HaveOccurred())
 		})
@@ -347,6 +347,203 @@ var _ = Describe("Route53 Resolver client", func() {
 				}
 			}
 			Expect(found).To(BeFalse())
+		})
+	})
+
+	When("deleting a hosted zone", func() {
+		var hostedZoneToFind *route53.CreateHostedZoneOutput
+
+		BeforeEach(func() {
+			now := time.Now()
+
+			hostedZoneToFind, err = rawRoute53Client.CreateHostedZone(&route53.CreateHostedZoneInput{
+				CallerReference: awssdk.String(fmt.Sprintf("1%d", now.UnixNano())),
+				Name:            awssdk.String("deleting.test.example.com"),
+			})
+			Expect(err).NotTo(HaveOccurred())
+		})
+
+		AfterEach(func() {
+			// We don't check error here because we actually expect an error since we just removed the hosted zone.
+			_, err = rawRoute53Client.DeleteHostedZoneWithContext(ctx, &route53.DeleteHostedZoneInput{Id: hostedZoneToFind.HostedZone.Id})
+		})
+
+		It("deletes the zone", func() {
+			err = route53Client.DeleteHostedZone(ctx, logger, *hostedZoneToFind.HostedZone.Id)
+			Expect(err).NotTo(HaveOccurred())
+
+			Eventually(func() (int, error) {
+				listHostedZoneResponse, err := rawRoute53Client.ListHostedZonesByName(&route53.ListHostedZonesByNameInput{
+					DNSName:  awssdk.String("deleting.test.example.com"),
+					MaxItems: awssdk.String("1"),
+				})
+				return len(listHostedZoneResponse.HostedZones), err
+			}, "3s", "100ms").Should(BeZero())
+		})
+	})
+
+	When("creating dns records", func() {
+		var hostedZoneToFind *route53.CreateHostedZoneOutput
+
+		BeforeEach(func() {
+			now := time.Now()
+
+			hostedZoneToFind, err = rawRoute53Client.CreateHostedZone(&route53.CreateHostedZoneInput{
+				CallerReference: awssdk.String(fmt.Sprintf("1%d", now.UnixNano())),
+				Name:            awssdk.String("dnsrecords.example.com"),
+			})
+			Expect(err).NotTo(HaveOccurred())
+		})
+
+		AfterEach(func() {
+			foundDnsRecordsResponse, err := rawRoute53Client.ListResourceRecordSetsWithContext(ctx, &route53.ListResourceRecordSetsInput{
+				HostedZoneId: hostedZoneToFind.HostedZone.Id,
+			})
+			Expect(err).NotTo(HaveOccurred())
+
+			for _, recordSet := range foundDnsRecordsResponse.ResourceRecordSets {
+				// We don't need to remove NS or SOA records.
+				if *recordSet.Type == "CNAME" || *recordSet.Type == "A" {
+					_, err = rawRoute53Client.ChangeResourceRecordSetsWithContext(ctx, &route53.ChangeResourceRecordSetsInput{
+						ChangeBatch: &route53.ChangeBatch{
+							Changes: []*route53.Change{
+								{
+									Action:            awssdk.String("DELETE"),
+									ResourceRecordSet: recordSet,
+								},
+							},
+						},
+						HostedZoneId: hostedZoneToFind.HostedZone.Id,
+					})
+					Expect(err).NotTo(HaveOccurred())
+				}
+			}
+
+			_, err = rawRoute53Client.DeleteHostedZoneWithContext(ctx, &route53.DeleteHostedZoneInput{Id: hostedZoneToFind.HostedZone.Id})
+			Expect(err).NotTo(HaveOccurred())
+		})
+
+		It("creates the records", func() {
+			dnsRecordsToCreate := []resolver.DNSRecord{
+				{
+					Kind:  "CNAME",
+					Name:  "a.dnsrecords.example.com",
+					Value: "something",
+				},
+				{
+					Kind:   "ALIAS",
+					Name:   "z.dnsrecords.example.com",
+					Value:  "alias-value",
+					Region: "eu-central-1",
+				},
+			}
+			err = route53Client.AddDnsRecordsToHostedZone(ctx, logger, *hostedZoneToFind.HostedZone.Id, dnsRecordsToCreate)
+			Expect(err).NotTo(HaveOccurred())
+
+			foundDnsRecordsResponse, err := rawRoute53Client.ListResourceRecordSetsWithContext(ctx, &route53.ListResourceRecordSetsInput{
+				HostedZoneId: hostedZoneToFind.HostedZone.Id,
+			})
+			Expect(err).NotTo(HaveOccurred())
+
+			cnameFound := false
+			aliasFound := false
+			for _, recordSet := range foundDnsRecordsResponse.ResourceRecordSets {
+				if *recordSet.Name == "a.dnsrecords.example.com." && *recordSet.ResourceRecords[0].Value == "something" {
+					cnameFound = true
+				}
+				if *recordSet.Name == "z.dnsrecords.example.com." && *recordSet.AliasTarget.DNSName == "alias-value" {
+					aliasFound = true
+				}
+			}
+			Expect(cnameFound).To(BeTrue())
+			Expect(aliasFound).To(BeTrue())
+
+			By("creating the exact same records again, it doesn't fail", func() {
+				dnsRecordsToCreate := []resolver.DNSRecord{
+					{
+						Kind:  "CNAME",
+						Name:  "a.dnsrecords.example.com",
+						Value: "something",
+					},
+					{
+						Kind:   "ALIAS",
+						Name:   "z.dnsrecords.example.com",
+						Value:  "alias-value",
+						Region: "eu-central-1",
+					},
+				}
+				err = route53Client.AddDnsRecordsToHostedZone(ctx, logger, *hostedZoneToFind.HostedZone.Id, dnsRecordsToCreate)
+				Expect(err).NotTo(HaveOccurred())
+			})
+		})
+	})
+	When("deleting all dns records from hosted zone", func() {
+		var hostedZoneToFind *route53.CreateHostedZoneOutput
+
+		BeforeEach(func() {
+			now := time.Now()
+
+			hostedZoneToFind, err = rawRoute53Client.CreateHostedZone(&route53.CreateHostedZoneInput{
+				CallerReference: awssdk.String(fmt.Sprintf("1%d", now.UnixNano())),
+				Name:            awssdk.String("dnsrecordstodelete.example.com"),
+			})
+			Expect(err).NotTo(HaveOccurred())
+
+			dnsRecordsToCreate := []resolver.DNSRecord{
+				{
+					Kind:  "CNAME",
+					Name:  "a.dnsrecordstodelete.example.com",
+					Value: "something",
+				},
+				{
+					Kind:   "ALIAS",
+					Name:   "z.dnsrecordstodelete.example.com",
+					Value:  "alias-value",
+					Region: "eu-central-1",
+				},
+			}
+			err = route53Client.AddDnsRecordsToHostedZone(ctx, logger, *hostedZoneToFind.HostedZone.Id, dnsRecordsToCreate)
+			Expect(err).NotTo(HaveOccurred())
+		})
+
+		AfterEach(func() {
+			foundDnsRecordsResponse, err := rawRoute53Client.ListResourceRecordSetsWithContext(ctx, &route53.ListResourceRecordSetsInput{
+				HostedZoneId: hostedZoneToFind.HostedZone.Id,
+			})
+			Expect(err).NotTo(HaveOccurred())
+
+			for _, recordSet := range foundDnsRecordsResponse.ResourceRecordSets {
+				// We don't need to remove NS or SOA records.
+				if *recordSet.Type == "CNAME" || *recordSet.Type == "A" {
+					_, err = rawRoute53Client.ChangeResourceRecordSetsWithContext(ctx, &route53.ChangeResourceRecordSetsInput{
+						ChangeBatch: &route53.ChangeBatch{
+							Changes: []*route53.Change{
+								{
+									Action:            awssdk.String("DELETE"),
+									ResourceRecordSet: recordSet,
+								},
+							},
+						},
+						HostedZoneId: hostedZoneToFind.HostedZone.Id,
+					})
+					Expect(err).NotTo(HaveOccurred())
+				}
+			}
+
+			_, err = rawRoute53Client.DeleteHostedZoneWithContext(ctx, &route53.DeleteHostedZoneInput{Id: hostedZoneToFind.HostedZone.Id})
+			Expect(err).NotTo(HaveOccurred())
+		})
+
+		It("removes all dns records", func() {
+			err = route53Client.DeleteDnsRecordsFromHostedZone(ctx, logger, *hostedZoneToFind.HostedZone.Id)
+			Expect(err).NotTo(HaveOccurred())
+
+			foundDnsRecordsResponse, err := rawRoute53Client.ListResourceRecordSetsWithContext(ctx, &route53.ListResourceRecordSetsInput{
+				HostedZoneId: hostedZoneToFind.HostedZone.Id,
+			})
+			Expect(err).NotTo(HaveOccurred())
+			// After removing the cluster dns records we expect 2 dns records: the NS and SOA records.
+			Expect(len(foundDnsRecordsResponse.ResourceRecordSets)).Should(BeNumerically("<=", 2))
 		})
 	})
 })